mk: Move REQUIRES generation into its own file.
This also fixes an issue seen with the generation of REQUIRES for gcc
packages.  The previous script didn't consider libraries inside DESTDIR
correctly, and for gcc ended up generating REQUIRES entries for libgcc
and libstdc++ using the system paths instead of ignoring them as they
are supplied by the package itself.